Replace sharescoreviewmodel functionality with GameStateManager
