+++ app/src/main/java/com/example/lostcitiesscorecalculator/ui/scoreboard/ScoreboardFragment.kt
@@ -25,8 +25,6 @@ class ScoreboardFragment : Fragment() {
     // onDestroyView.
     private val binding get() = _binding!!
 
-    private lateinit var sharedScoreViewModel: SharedScoreViewModel
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -50,13 +48,12 @@ class ScoreboardFragment : Fragment() {
             onEndGameButtonPressed()
         }
 
-        sharedScoreViewModel = (requireActivity().application as LostCitiesScoreCalculatorApplication).sharedScoreViewModel
-
-        sharedScoreViewModel.player1CurrentPoints.observe(viewLifecycleOwner, player1CurrentScoreObserver)
-        sharedScoreViewModel.player2CurrentPoints.observe(viewLifecycleOwner, player2CurrentScoreObserver)
-        sharedScoreViewModel.player1TotalPoints.observe(viewLifecycleOwner, player1TotalScoreObserver)
-        sharedScoreViewModel.player2TotalPoints.observe(viewLifecycleOwner, player2TotalScoreObserver)
-        sharedScoreViewModel.roundScores.observe(viewLifecycleOwner, roundScoreObserver)
+        GameStateManager.player1CurrentPoints.observe(viewLifecycleOwner, player1CurrentScoreObserver)
+        GameStateManager.player2CurrentPoints.observe(viewLifecycleOwner, player2CurrentScoreObserver)
+        GameStateManager.player1TotalPoints.observe(viewLifecycleOwner, player1TotalScoreObserver)
+        GameStateManager.player2TotalPoints.observe(viewLifecycleOwner, player2TotalScoreObserver)
+        GameStateManager.roundScores.observe(viewLifecycleOwner, roundScoreObserver)
+        GameStateManager.roundScores.observe(viewLifecycleOwner, roundScoreObserver)
 
         GameStateManager.player1Name.observe(viewLifecycleOwner, player1NameObserver)
         GameStateManager.player2Name.observe(viewLifecycleOwner, player2NameObserver)
